LPD-61412 Split into components
