+++ modules/apps/staging/staging-taglib/src/main/resources/META-INF/resources/js/index.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {ExportErrorsReport} from './components/ExportErrorsReport';
+export {ExportErrorsReportAction} from './components/ExportErrorsReportAction';
 
 // @ts-ignore
