+++ modules/apps/staging/staging-taglib/src/main/resources/META-INF/resources/process_list_menu/items/details.jspf
@@ -22,7 +22,7 @@
 		</button>
 
 		<react:component
-			module="{ExportErrorsReport} from staging-taglib"
+			module="{ExportErrorsReportAction} from staging-taglib"
 		/>
 	</li>
 </c:if>
